fix: exposing multiple modules from project-b
