+++ packages/examples/project-b/federation.config.js
@@ -2,7 +2,8 @@ export const federationconfig = () => ({
   name: 'sample_project_b',
   filename: 'my-remote-entry.js',
   exposes: {
-    './button': './src/index.js',
+    './button': './src/button.js',
+    './link': './src/link.js',
   },
   shared: {
     react: {
