+++ packages/examples/project-b/src/button.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import axios from 'axios';
+
+const Button = (text) => ({
+  text,
+});
+
+export const someThingElse = () => ({
+  something: 'nothing',
+  hello: 'world',
+  React,
+  axios,
+});
+
+export default Button;
